requirements.txt
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
blinker==1.5
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==40.0.1
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
Flask==2.2.3
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kaitaistruct==0.10
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pyee==9.0.4
Pygments==2.14.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
stack-data==0.6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
webdriver-manager==3.8.5
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
zstandard==0.20.0
